# PERF: MultiIndex.get_locs