@jreback 2022-02-11
this fully covers https://github.com/pandas-dev/pandas/issues/38650 ?