@jreback 2022-02-11
thanks @lukemanley 

ok to leave that issue open though this seems to get a lot of the low hanging fruit so maybe not worth it but happy to be proven wrong