@lukemanley 2022-02-11
>ok to leave that issue open

sound good. Its now open