@rhshadrach 2022-02-19
This looks to have a great performance improvement on partial indexing, which I think should also get a line in the whatsnew.